Rename base_auth
