+++ src/middleware/auth/base-auth.ts
